+++ src/main/java/com/backend/Domain/User/UserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findById(Integer Id);
+    Optional<User> findById(Long Id);
 
     Optional<User> findByEmail(String email);
 }
